+++ CardinalKit-Example/CardinalKit/Components/Onboarding/PageViewController.swift
@@ -1,0 +1,127 @@
+//
+//  PageViewController.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import UIKit
+import SwiftUI
+
+struct PageControl: UIViewRepresentable {
+    var numberOfPages: Int
+    @Binding var currentPage: Int
+    func makeCoordinator() -> Coordinator {
+        Coordinator(self)
+    }
+
+    func makeUIView(context: Context) -> UIPageControl {
+        let control = UIPageControl()
+        let config = CKPropertyReader(file: "CKConfiguration")
+        control.numberOfPages = numberOfPages
+        control.pageIndicatorTintColor = UIColor.lightGray
+        control.currentPageIndicatorTintColor = config.readColor(query: "Primary Color")
+        control.addTarget(
+            context.coordinator,
+            action: #selector(Coordinator.updateCurrentPage(sender:)),
+            for: .valueChanged)
+
+        return control
+    }
+
+    func updateUIView(_ uiView: UIPageControl, context: Context) {
+        uiView.currentPage = currentPage
+    }
+
+    class Coordinator: NSObject {
+        var control: PageControl
+
+        init(_ control: PageControl) {
+            self.control = control
+        }
+        @objc
+        func updateCurrentPage(sender: UIPageControl) {
+            control.currentPage = sender.currentPage
+        }
+    }
+}
+
+struct PageView<Page: View>: View {
+    var viewControllers: [UIHostingController<Page>]
+    @State var currentPage = 0
+    init(_ views: [Page]) {
+        self.viewControllers = views.map { UIHostingController(rootView: $0) }
+    }
+
+    var body: some View {
+        ZStack(alignment: .bottom) {
+            PageViewController(controllers: viewControllers, currentPage: $currentPage)
+            PageControl(numberOfPages: viewControllers.count, currentPage: $currentPage)
+        }
+    }
+}
+
+struct PageViewController: UIViewControllerRepresentable {
+    var controllers: [UIViewController]
+    @Binding var currentPage: Int
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(self)
+    }
+
+    func makeUIViewController(context: Context) -> UIPageViewController {
+        let pageViewController = UIPageViewController(
+            transitionStyle: .scroll,
+            navigationOrientation: .horizontal)
+        pageViewController.dataSource = context.coordinator
+        pageViewController.delegate = context.coordinator
+
+        return pageViewController
+    }
+
+    func updateUIViewController(_ pageViewController: UIPageViewController, context: Context) {
+        pageViewController.setViewControllers(
+        [self.controllers[self.currentPage]], direction: .forward, animated: true)
+    }
+
+    class Coordinator: NSObject, UIPageViewControllerDataSource, UIPageViewControllerDelegate {
+        var parent: PageViewController
+
+        init(_ pageViewController: PageViewController) {
+            self.parent = pageViewController
+        }
+
+        func pageViewController(
+            _ pageViewController: UIPageViewController,
+            viewControllerBefore viewController: UIViewController) -> UIViewController? {
+            guard let index = parent.controllers.firstIndex(of: viewController) else {
+                return nil
+            }
+            if index == 0 {
+                return parent.controllers.last
+            }
+            return parent.controllers[index - 1]
+        }
+
+        func pageViewController(
+            _ pageViewController: UIPageViewController,
+            viewControllerAfter viewController: UIViewController) -> UIViewController? {
+            guard let index = parent.controllers.firstIndex(of: viewController) else {
+                return nil
+            }
+            if index + 1 == parent.controllers.count {
+                return parent.controllers.first
+            }
+            return parent.controllers[index + 1]
+        }
+
+        func pageViewController(_ pageViewController: UIPageViewController, didFinishAnimating finished: Bool, previousViewControllers: [UIViewController], transitionCompleted completed: Bool) {
+            if completed,
+                let visibleViewController = pageViewController.viewControllers?.first,
+                let index = parent.controllers.firstIndex(of: visibleViewController) {
+                parent.currentPage = index
+            }
+        }
+    }
+}
