+++ CardinalKit-Example/CardinalKit/Components/Profile/ReportView.swift
@@ -1,0 +1,32 @@
+//
+//  ReportView.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import SwiftUI
+
+struct ReportView: View {
+    let color: Color
+    var email = ""
+    
+    init(color: Color, email: String) {
+        self.color = color
+        self.email = email
+    }
+    
+    var body: some View {
+        HStack {
+            Text("Report a Problem")
+            Spacer()
+            Text(self.email).foregroundColor(self.color)
+        }
+        .frame(height: 60)
+        .contentShape(Rectangle())
+        .gesture(TapGesture().onEnded({
+            EmailHelper.shared.sendEmail(subject: "App Support Request", body: "Enter your support request here.", to: self.email)
+        }))
+    }
+}
