+++ CardinalKit-Example/CardinalKit/Components/Profile/SupportView.swift
@@ -1,0 +1,35 @@
+//
+//  SupportView.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import SwiftUI
+
+struct SupportView: View {
+    let color: Color
+    var phone = ""
+    
+    init(color: Color, phone: String) {
+        self.color = color
+        self.phone = phone
+    }
+    
+    var body: some View {
+        HStack {
+            Text("Support")
+            Spacer()
+            Text(self.phone).foregroundColor(self.color)
+        }
+        .frame(height: 60)
+        .contentShape(Rectangle())
+        .gesture(TapGesture().onEnded({
+            let telephone = "tel://"
+                let formattedString = telephone + self.phone
+            guard let url = URL(string: formattedString) else { return }
+            UIApplication.shared.open(url)
+        }))
+    }
+}
