+++ CardinalKit-Example/CardinalKit/Components/Profile/WithdrawView.swift
@@ -1,0 +1,34 @@
+//
+//  WithdrawView.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import SwiftUI
+
+struct WithdrawView: View {
+    let color: Color
+    @State var showWithdraw = false
+    
+    init(color: Color) {
+        self.color = color
+    }
+    
+    var body: some View {
+        HStack {
+            Text("Withdraw from Study").foregroundColor(self.color)
+            Spacer()
+            Text("›").foregroundColor(self.color)
+        }.frame(height: 60)
+            .contentShape(Rectangle())
+            .gesture(TapGesture().onEnded({
+                self.showWithdraw.toggle()
+            })).sheet(isPresented: $showWithdraw, onDismiss: {
+                
+            }, content: {
+                WithdrawalViewController()
+            })
+    }
+}
