+++ CardinalKit-Example/CardinalKit/Components/Profile/WithdrawalViewController.swift
@@ -1,0 +1,80 @@
+//
+//  WithdrawalViewController.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import UIKit
+import SwiftUI
+import ResearchKit
+import Firebase
+import CardinalKit
+
+struct WithdrawalViewController: UIViewControllerRepresentable {
+    
+    func makeCoordinator() -> Coordinator {
+        Coordinator()
+    }
+
+    typealias UIViewControllerType = ORKTaskViewController
+    
+    func updateUIViewController(_ taskViewController: ORKTaskViewController, context: Context) {}
+    func makeUIViewController(context: Context) -> ORKTaskViewController {
+
+        let config = CKPropertyReader(file: "CKConfiguration")
+        
+        let instructionStep = ORKInstructionStep(identifier: "WithdrawlInstruction")
+        instructionStep.title = NSLocalizedString(config.read(query: "Withdrawal Instruction Title"), comment: "")
+        instructionStep.text = NSLocalizedString(config.read(query: "Withdrawal Instruction Text"), comment: "")
+        
+        let completionStep = ORKCompletionStep(identifier: "Withdraw")
+        completionStep.title = NSLocalizedString(config.read(query: "Withdraw Title"), comment: "")
+        completionStep.text = NSLocalizedString(config.read(query: "Withdraw Text"), comment: "")
+        
+        let withdrawTask = ORKOrderedTask(identifier: "Withdraw", steps: [instructionStep, completionStep])
+        
+        // wrap that task on a view controller
+        let taskViewController = ORKTaskViewController(task: withdrawTask, taskRun: nil)
+        
+        taskViewController.delegate = context.coordinator // enables `ORKTaskViewControllerDelegate` below
+        
+        // & present the VC!
+        return taskViewController
+
+    }
+
+    class Coordinator: NSObject, ORKTaskViewControllerDelegate {
+        public func taskViewController(_ taskViewController: ORKTaskViewController, didFinishWith reason: ORKTaskViewControllerFinishReason, error: Error?) {
+            switch reason {
+            case .completed:
+                UserDefaults.standard.set(false, forKey: "didCompleteOnboarding")
+                
+                do {
+                    try Auth.auth().signOut()
+                    
+                    if (ORKPasscodeViewController.isPasscodeStoredInKeychain()) {
+                        ORKPasscodeViewController.removePasscodeFromKeychain()
+                    }
+                    
+                    taskViewController.dismiss(animated: true, completion: {
+                        //TODO: recover app without error
+                        fatalError()
+                    })
+                    
+                } catch {
+                    print(error.localizedDescription)
+                    Alerts.showInfo(title: "Error", message: error.localizedDescription)
+                }
+                
+            default:
+                
+                // otherwise dismiss onboarding without proceeding.
+                taskViewController.dismiss(animated: true, completion: nil)
+                
+            }
+        }
+    }
+    
+}
