+++ CardinalKit-Example/CardinalKit/Components/__LaunchContainerViewController.swift
@@ -20,6 +20,7 @@ class LaunchContainerViewController: UIViewController {
     }
 
     override func viewDidLoad() {
+        print("LaunchContainerViewController viewDidLoad")
         super.viewDidLoad()
         // Do any additional setup after loading the view.
