+++ CardinalKit-Example/CardinalKit/Supporting Files/CKConfiguration-new.plist
@@ -42,7 +42,7 @@
 	<string>On the next screen, you will be prompted to grant access to read and write some of your general and health information, such as height, weight, and steps taken so you don&apos;t have to enter it again.</string>
 	<key>Background Read Frequency</key>
 	<string>immediate</string>
-	<key>Sign in with Apple</key>
+	<key>Login-Sign-In-With-Apple</key>
 	<dict>
 		<key>Enabled</key>
 		<true/>
