+++ CardinalKit-Example/CardinalKit/Utils/Extensions+UIColor.swift
@@ -1,0 +1,19 @@
+//
+//  Extensions+UIColor.swift
+//  CardinalKit_Example
+//
+//  Created by Santiago Gutierrez on 10/12/20.
+//  Copyright © 2020 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+public extension UIColor {
+    class func greyText() -> UIColor {
+        return UIColor(netHex: 0x989998)
+    }
+    
+    class func lightWhite() -> UIColor {
+        return UIColor(netHex: 0xf7f8f7)
+    }
+}
